Quit with same Component for logic reuse
